Refactor: Code splitting
- Separate flashcards hooks for readabilities.
- Update web/README.md
- Remove unnecessary css class (use CSS classes directly in component)
- Remove utilities/getProtocol.js
