+++ web/lib/components/flashcards/style.module.css
@@ -1,7 +1,3 @@
-.layout {
-  @apply container;
-}
-
 .flipcard {
   @apply m-auto aspect-video w-9/12 bg-transparent;
   perspective: 1000px;
@@ -26,28 +22,12 @@
 }
 
 .swiper {
-  width: 100%;
-  height: 100%;
+  @apply h-full w-full;
 }
 
 .swiper-slide {
-  text-align: center;
+  @apply flex items-center justify-center bg-white text-center;
   font-size: 18px;
-  background: #fff;
-
-  /* Center slide text vertically */
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: -webkit-flex;
-  display: flex;
-  -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  -webkit-justify-content: center;
-  justify-content: center;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  -webkit-align-items: center;
-  align-items: center;
 }
 
 .swiper-horizontal > .swiper-pagination-bullets,
