--- web/lib/utilities/getProtocol.js
@@ -1,3 +1,0 @@
-export function getProtocol() {
-  return process.env.NODE_ENV === 'development' ? 'http://' : 'https://'
-}
